backend/digi_yatra/requirements.txt
asgiref>=3.8.1,<4.0
click>=8.1.8,<9.0
Django>=5.1.4,<6.0
django-cors-headers>=4.6.0,<5.0
dlib==19.22.0  # Stable version of dlib that may not require compilation
face-recognition>=1.3.0,<2.0
numpy>=1.26.0,<2.0
pillow>=11.0.0,<12.0
sqlparse>=0.5.3,<0.6.0
uv>=0.5.4,<1.0
